+++ packages/backend/index.ts
@@ -1,26 +1,14 @@
-import Fastify from "fastify";
+import fastify from "fastify";
 import fastifyStatic from "fastify-static";
 import mercurius from "mercurius";
+import { schema } from "@packages/schema";
 import { config } from "@packages/config";
 
-export const app = Fastify();
-
-const schema = `
-type Query {
-  hello: String!
-}
-`;
-
-const resolvers = {
-  Query: {
-    hello: () => `Hello World !`,
-  },
-};
+export const app = fastify();
 
 // Mercurius GraphQL server
 app.register(mercurius, {
   schema,
-  resolvers,
   graphiql: config.mode === "development" && "playground",
 });